# Change alias models to use smaller vertex footprint