--- a/source/psp/gu_vertex_lighting.h
+++ b/source/psp/gu_vertex_lighting.h
@@ -10,4 +10,5 @@ extern byte anorm_pitch[162];
 extern byte anorm_yaw[162];
 extern byte vlighttable[256][256];
 
+float VLight_GetLightValue(int index, float apitch, float ayaw);
 float VLight_LerpLight(int index1, int index2, float ilerp, float apitch, float ayaw);