These changes have no perceptible visual changes.

UV coords go from 32f to s16, in theory still room to go lower, but because of 32bit alignment for verts, it's hard to get any benefit from it
XYZ coords go from 32f to s8, matches precision that mdl stores them in anyway. However PSP interprets the signed -128 to 127 range as -1 to 1, so scaling had to be changed.

These changes yield a tiny boost in framerates in timedemos and a smaller command buffer alias models in the hunk.
```
before
timedemo 1:
12.8s 75.5fps
timedemo 2:
11.2s 88fps
timedemo 3:
16.3s 66.7fps

after
timedemo 1:
12.4s 78fps
timedemo 2:
11.0s 89fps
timedemo 3:
15.6s 69.7fps
```

In order to boost model performance more, vertex lighting should be replaced with sceGuColor (HIGHLY recommended for nzp), which cuts vertex size by yet another 32 bits, and removes a lot of the per vertex calculation, with frankly minimal visual impact. additionally it could save you some much needed ram, which could buy you opportunities to utilize command buffer for positions too